refactor: 온보딩을 한 페이지에서 관리하도록 변경, 질문 내용 상수화 (#4)
